+++ Magic/CHANGELOG.md
@@ -68,6 +68,7 @@
  - Spell parameters can now be used inside of equations within that spell
  - The `ModifyAttribute` action now supports an `x` variable holding the current attribute value, just like `ModifyProperties`
  - The `/mage deactivate` command can now be used as an alias for `/mage lock`
+ - Add support for JSON chat component messages
 
 ## Changes